test fixes

+++ hadoop-tools/hadoop-posum/src/main/java/org/apache/hadoop/tools/posum/simulation/predictor/basic/BasicPredictionStats.java
@@ -57,8 +57,8 @@ public class BasicPredictionStats extends PredictionStats {
   }
 
   public void addSource(JobProfile job) {
-    avgMapDuration = addValue(job.getAvgMapDuration().doubleValue(), avgMapDuration, mapDurations);
-    avgReduceDuration = addValue(job.getAvgReduceDuration().doubleValue(), avgReduceDuration, reduceDurations);
+    avgMapDuration = addValue(job.getAvgMapDuration(), avgMapDuration, mapDurations);
+    avgReduceDuration = addValue(job.getAvgReduceDuration(), avgReduceDuration, reduceDurations);
     incrementSampleSize();
   }
